Refactor route polyline rendering: extract step color and dash pattern logic into utility functions and update travel mode type
