+++ mobile/src/components/LocationScreen/RoutePolyline.tsx
@@ -1,80 +1,14 @@
 import { Polyline } from "react-native-maps";
 import { COLORS } from "../../constants";
+import { RouteInfo, TRAVEL_MODE, TravelMode } from "../../types/Directions";
 import {
-  RouteInfo,
-  StepInfo,
-  TRAVEL_MODE,
-  TravelMode,
-} from "../../types/Directions";
+  getStepColor,
+  getStepLineDashPattern,
+} from "../../utils/pathTypeUtils";
 
 interface RoutePolylineProps {
   readonly route: RouteInfo | null;
   readonly travelMode: TravelMode | null;
-}
-
-/**
- * Determines the color for a step based on vehicle type or travel mode.
- */
-function getStepColor(step: StepInfo): string {
-  // If it's a transit step with vehicle details, color by vehicle type
-  if (step.transitDetails) {
-    const vehicleType = step.transitDetails.vehicleType?.toUpperCase();
-    switch (vehicleType) {
-      case "SUBWAY":
-      case "RAIL":
-        return "#FF6B35"; // Metro/Rail: Orange-red
-      case "BUS":
-        return "#1E90FF"; // Bus: Light blue
-      case "TROLLEYBUS":
-        return "#4169E1"; // Trolleybus: Royal blue
-      default:
-        return "#1E90FF"; // Default transit: Light blue
-    }
-  }
-
-  // Check travel mode for non-transit steps
-  if (step.travelMode) {
-    const mode = step.travelMode.toUpperCase();
-    switch (mode) {
-      case "BICYCLE":
-        return "#FF8C00"; // Cycling: Orange
-      case "DRIVE":
-        return "#40509F"; // Driving: Dark blue
-      case "WALK":
-      default:
-        return "#40509F"; // Walking: Dark blue
-    }
-  }
-
-  // Default: Dark blue for walking
-  return "#40509F";
-}
-
-/**
- * Determines the line dash pattern for a step.
- */
-function getStepLineDashPattern(step: StepInfo): number[] | undefined {
-  // If it's a transit step, use solid line
-  if (step.transitDetails) {
-    return undefined;
-  }
-
-  // Check travel mode for non-transit steps
-  if (step.travelMode) {
-    const mode = step.travelMode.toUpperCase();
-    switch (mode) {
-      case "BICYCLE":
-        return [12, 6]; // Cycling: Dashed line
-      case "WALK":
-        return [8, 6]; // Walking: Dotted line
-      case "DRIVE":
-      default:
-        return undefined; // Driving: Solid line
-    }
-  }
-
-  // Default: Dotted line for walking
-  return [8, 6];
 }
 
 /**
