+++ mobile/src/services/DirectionsService.ts
@@ -91,7 +91,7 @@ function parseRoute(data: DirectionsResponse): RouteInfo | null {
         coordinates: step.polyline?.encodedPolyline
           ? decodePolyline(step.polyline.encodedPolyline)
           : [],
-        travelMode: step.travelMode, // Preserve travel mode for step-level coloring
+        travelMode: step.travelMode as TravelMode, // Preserve travel mode for step-level coloring
       };
 
       // Parse transit details if present
